+++ commands/Musica/play.js
@@ -1,8 +1,0 @@
-module.exports = {
-    name: 'play',
-    description: 'Posa una cançó al canal.',
-    type: 'musica',
-    execute(message) {
-        return message.reply("La funció de música està en desenvolupament...");
-    },
-};